Add Rails LoggedTagging support
Add compatibility for tags created for use with LoggedTagging. Since this is a using json the
processing is quite different.

Tags are added to the JSON in a tag field if it is only a string or proc or a key and value if it is
a symbol. (Maybe we will check the return value for procs in future to see if it is a structure with
keys/values but for the moment lets keep it simple and compatible)