Fix JRuby compatibility issue
Fix a JRuby compatibility issue and add jruby-head (1.7) to the travis
configuration. While fixing this issue another issue popped up with
1.6.7 compabitility. Not sure how to fix it yet so lets get a failure
on Travis which we can point JRuby people to.

The fact that the failing test is one that comes unmodified from the
rails test suite (which I assume is passing) makes me wonder if some
setup is missing on my part for this test related to JRuby.